requirements.txt
# Core Django
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-extensions==3.2.3
drf-spectacular==0.27.2
django-ratelimit==4.1.0
django-redis==5.4.0

# Database
psycopg2-binary==2.9.9
# pgbouncer==1.23.1

# Cache & Async
celery==5.4.0
redis==5.0.8
kombu==5.3.4

# AI / Deep Learning
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==2.4.1
# torchaudio==2.4.1
# transformers==4.44.2
# openai-whisper==20240930
faster-whisper==1.0.3
python-Levenshtein==0.25.1
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
# phonemizer==3.2.1
jiwer==3.0.4  # Word Error Rate

# File & Storage
boto3==1.35.13
pillow==10.4.0
python-magic==0.4.27
pydub==0.25.1
ffmpeg-python==0.2.0

# Monitoring & Logging
sentry-sdk==2.14.0
prometheus-client==0.21.0

# Utils
python-decouple==3.8
gunicorn==23.0.0
uvicorn==0.30.6
httpx==0.27.2
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
faker==28.4.1
coverage==7.6.1
flake8==7.1.1
black==24.8.0
